# reduce the size of Attr from 3 pointers to 2 on 64 bit machines