--- a/src/libexpr/function-trace.hh
+++ b/src/libexpr/function-trace.hh
@@ -8,7 +8,7 @@ namespace nix {
 
 struct FunctionCallTrace
 {
-    const Pos & pos;
+    const Pos pos;
     FunctionCallTrace(const Pos & pos);
     ~FunctionCallTrace();
 };