… src/libexpr/flake/flake.cc
… src/libexpr/lexer.l
… src/libexpr/nixexpr.cc
… src/libexpr/parser.y
… src/libexpr/primops.cc
… src/libexpr/primops.hh
… src/libexpr/primops/context.cc
… src/libexpr/primops/fetchClosure.cc
… src/libexpr/primops/fetchMercurial.cc
… src/libexpr/primops/fetchTree.cc
… src/libexpr/primops/fromTOML.cc
… src/libexpr/symbol-table.hh
… src/libexpr/value-to-json.cc
… src/libexpr/value-to-json.hh
… src/libexpr/value-to-xml.cc
… src/libexpr/value-to-xml.hh
… src/libexpr/value.hh
… src/libutil/chunked-vector.hh
… src/libutil/error.hh
… src/libutil/ref.hh
… src/libutil/tests/chunked-vector.cc
… src/libutil/types.hh
… src/nix-env/nix-env.cc
… src/nix-env/user-env.cc
… src/nix-instantiate/nix-instantiate.cc
… src/nix/app.cc
… src/nix/bundle.cc
… src/nix/edit.cc
… src/nix/eval.cc
… src/nix/flake.cc