this changes position handling to use a table similar to SymbolTable for storing position information, and changes SymbolTable itself to require less memory per Symbol at the expense of lookup convenience. with both of these tables using 32 bit indices we can pack Attr more tightly, reducing GC heap size and allocations by 10-15%. reduced GC load also gives us an eval performance improvement on the same order for free.

depends on #5865