@pennae 2022-03-11
puzzling that the `eval-okay-xml` test fails on macos, it runs fine here on linux :confused: