@tomberek 2022-03-23
NIX_SHOW_STATS trimmed for brevity
```
before                                             after
{                                                  {
  "envs": {                                          "envs": {
    "bytes": 146264296                                 "bytes": 146264296
  "list": {                                          "list": {
    "bytes": 26839376,                                 "bytes": 26839352,
  "values": {                                        "values": {
    "bytes": 362397240                                 "bytes": 362397456
  "symbols": {                                       "symbols": {
    "bytes": 3618409                                   "bytes": 1382817
  "sets": {                                          "sets": {
    "bytes": 1100289680,                               "bytes": 747502224,
  },                                                 },
  "sizes": {                                         "sizes": {
    "Env": 16,                                         "Env": 16,
    "Value": 24,                                       "Value": 24,
    "Bindings": 16,                                    "Bindings": 16,
    "Attr": 24                                         "Attr": 16
  },                                                 },
  "gc": {                                            "gc": {
    "heapSize": 1561522176,                            "heapSize": 1158864896,
    "totalBytes": 1960805392                           "totalBytes": 1566740096
  }                                                  }
```