@pennae 2022-03-29
rebased again. also renamed `PosPtr` to `PosIdx` and `SymbolPtr` to `SymbolIdx` since neither of them really act like pointers at all.