@Ericson2314 2022-03-31
I am not up to date on these things and so could not review very way, but perhaps @layus can after https://github.com/NixOS/nix/pull/6204. I think it's good to have the people that do performance and diagnostic quality reviewing each other, as those goals in the general case have some tension.

(To be clear, I don't see a sign of diagnostic quality regression in this case.)