@pennae 2022-04-18
haven't tested 32 bit, but it should also decrease. Pos now elides a Symbol (aka pointer) per instances at the cost of two ints and a std::string per file. that should break even at five positions per file on average for 32 bit machines, and if nixpkgs is anything to go by that shouldn't be hard to achieve :)