@pennae 2022-04-22
not having high hopes for CI to be honest, the installer test has been hating this PR for weeks xD