@pennae 2022-04-22
looks like the auto-merge was a bit over-eager again, the macos build failed:

```
 nix> In file included from src/libutil/tests/chunked-vector.cc:1:
nix> src/libutil/chunked-vector.hh:3:10: fatal error: 'bits/stdint-uintn.h' file not found
nix> #include <bits/stdint-uintn.h>
nix>          ^~~~~~~~~~~~~~~~~~~~~
nix> 1 error generated.
```