@dtzWill 2022-04-22
Hi, I was wondering if `std::deque` was considered/would work instead of ChunkedVector? deque is basically a vector of vectors and is useful for same reasons I think ChunkedVector was introduced?

(modulo the forEach helper, but `std::for_each` will do the trick)