@pennae 2022-04-22
@dtzWill was considered but (maybe mistakenly) deemed unsuitable because there are no allocation size guarantees for its chunks. part of the reason for chunking is to keep a constant maximum memory overhead for unused capacity, if we just didn't find the way to get that done with a deque then a deque would work just as well