@edolstra 2022-04-22
Question: we now have two ways to uniquely refer to symbols, namely `Symbol` and `SymbolIdx`, which is confusing. Which one should be used in APIs? Given that it's possible to efficiently convert a `SymbolIdx` to a `Symbol` but not the other way around, does it even make sense to expose `Symbol`?