@roberth 2023-04-12
@pennae what do you think of a scheme like this, that replaces the position table by a clever indexing scheme?
 - https://github.com/NixOS/nix/issues/8208

It would enable some improvements to error messages; see the github reverse links to the issues that come after it.